# RESTApiDesign/openapi_todomanager_2.0.0.yaml
openapi: 3.0.1
info:
  title: ToDo Manager
  description: DSP-Lab-2
  contact:
    email: ricioppo.guido@gmail.com
  license:
    name: GPL 2.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  version: 1.0.0
...
servers:
- url: https://todomanager.dsp.io/v1
- url: http://todomanager.dsp.io/v1
...
paths:
...
  /api/tasks/{id}:
    get:
      tags:
      - task
      summary: Find task by ID
      description: Returns a single task
      operationId: getTaskById
      parameters:
      - name: id
        in: path
        description: ID of a task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "400":
          description: Invalid ID supplied
          content: {}
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Task
    put:
      tags:
      - task
      summary: Updates a task
      operationId: updateTask
      parameters:
      - name: id
        in: path
        description: ID of a task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
                description: The reason for the bad request
                x-content-type: text/plain
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Task
    delete:
      tags:
      - task
      summary: Deletes a task
      operationId: deleteTask
      parameters:
      - name: id
        in: path
        description: ID of a task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Task
    patch:
      tags:
      - task
      summary: Updates "completed" field of a task
      operationId: updateCompletedTask
      parameters:
      - name: id
        in: path
        description: ID of a task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Task
...
  /api/users:
    get:
      tags:
      - users
      summary: Get information about the users
      description: Retrieve the available information (passwords excluded) about all
        the users.
      operationId: getAllUsers
      parameters:
      - name: pageNumber
        in: query
        description: Page number of users list
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Users
  /api/users/{userId}:
    get:
      tags:
      - user
      summary: Get information about a user
      description: Get user resource (password excluded) identified by userId
      operationId: getUserById
      parameters:
      - name: userId
        in: path
        description: ID of the user to get
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid ID supplied
          content: {}
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: User
...
components:
  schemas:
    Project:
      type: string
      enum:
      - Personal
      - WA1_Project
      - WA2_Project
      - DSP_Project
    Task:
      required:
      - description
      type: object
      properties:
        id:
          type: integer
          description: unique identifier for the task
        description:
          maxLength: 160
          type: string
        important:
          type: boolean
          default: false
        private:
          type: boolean
          description: "True if only the user who created it can get information about\
            \ it, False if every user can get information about it"
          default: true
        project:
          $ref: '#/components/schemas/Project'
        deadline:
          pattern: "^(2[1-9][0-9]{2}|20[2-9][0-9]|[3-9][0-9]{3})"
          type: string
          format: date-time
        completed:
          type: boolean
          default: false
        owner:
          type: string
          format: uri
        assignedTo:
          type: array
          items:
            type: string
            format: uri
        self:
          type: string
          format: uri
...
    User:
      required:
      - email
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        password:
          maxLength: 20
          minLength: 6
          type: string
        self:
          type: string
          format: uri
      example:
        password: password
        name: name
        self: http://example.com/aeiou
        id: 1
        email: email
    Users:
      type: object
      properties:
        totalPages:
          type: integer
        pageNumber:
          type: integer
        pageItems:
          type: array
          items:
            $ref: '#/components/schemas/User'
        next:
          type: string
          format: uri
        self:
          type: string
          format: uri
      example:
        next: http://example.com/aeiou
        pageNumber: 6
        totalPages: 0
        pageItems:
        - password: password
          name: name
          self: http://example.com/aeiou
          id: 1
          email: email
        - password: password
          name: name
          self: http://example.com/aeiou
          id: 1
          email: email
        self: http://example.com/aeiou
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT identifying user session
      name: JWT.token
      in: cookie